Uses `size_t` and adds aliases for clarity.

+++ src/emulator/system/registers.h
@@ -10,6 +10,7 @@
 
 #include <stdint.h>
 #include <stdbool.h>
+#include <stddef.h>
 
 /// A struct representing the PSTATE fields of the processor.
 typedef struct {
@@ -63,7 +64,7 @@ Registers createReg(void);
 
 static void initReg(Registers *reg);
 
-uint64_t getReg(Registers *reg, uint8_t id);
+uint64_t getReg(Registers *reg, size_t id);
 
 uint64_t getRegZR(Registers *reg);
 
@@ -73,9 +74,9 @@ uint64_t getRegSP(Registers *reg);
 
 bool getRegState(Registers *reg, PStateField field);
 
-void setReg64(Registers *reg, uint8_t id, uint64_t value);
+void setReg64(Registers *reg, size_t id, uint64_t value);
 
-void setReg32(Registers *reg, uint8_t id, uint32_t value);
+void setReg32(Registers *reg, size_t id, uint32_t value);
 
 void setRegZR64(__attribute__((unused)) Registers *reg, __attribute__((unused)) uint64_t value);
